move core api and flow implementations to a layer above for sharing
