+++ code/client/webpack.config.js
@@ -86,9 +86,9 @@ module.exports = {
   plugins: [
     new webpack.EnvironmentPlugin({
       PUBLIC_URL: '',
-      REACT_APP_NETWORK: '',
-      REACT_APP_RELAYER: '',
-      REACT_APP_RELAYER_SECRET: 'onewallet',
+      NETWORK: '',
+      RELAYER: '',
+      RELAYER_SECRET: 'onewallet',
       DEBUG: false,
       MIN_WALLET_VERSION: 0,
       SENTRY_DSN: '', // dev
