+++ code/lib/api/message.js
@@ -1,0 +1,10 @@
+let message
+const setMessage = (newMessage) => {
+  message = newMessage
+}
+const getMessage = () => message
+
+module.exports = {
+  setMessage,
+  getMessage
+}
