+++ code/lib/api/storage.js
@@ -1,0 +1,10 @@
+let storage
+const setStorage = (newStorage) => {
+  storage = newStorage
+}
+const getStorage = () => storage
+
+module.exports = {
+  setStorage,
+  getStorage
+}
