+++ code/package.json
@@ -18,7 +18,8 @@
     "hi-base32": "^0.5.1",
     "jssha": "^3.2.0",
     "keccak": "^3.0.1",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "web3": "^1.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.14.2",
